# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Cafetería",
        "description": "API para gestión de cafetería",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000",
            "description": "Servidor Local"
        }
    ],
    "paths": {
...
        "/api/sesiones": {
            "get": {
                "tags": [
                    "Sesiones"
                ],
                "summary": "Obtener lista de sesiones",
                "operationId": "9b4b4d87468265ce5c414289e2fc797b",
                "responses": {
                    "200": {
                        "description": "Lista completa de sesiones"
                    }
                }
            },
            "post": {
                "tags": [
                    "Sesiones"
                ],
                "summary": "Crear una nueva sesión",
                "operationId": "7fe5593f14e25be3bc468d12d0b17776",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "usuario_id",
                                    "ultima_actividad"
                                ],
                                "properties": {
                                    "usuario_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "agente": {
                                        "type": "string",
                                        "example": "Mozilla/5.0"
                                    },
                                    "ip": {
                                        "type": "string",
                                        "format": "ipv4",
                                        "example": "192.168.1.1"
                                    },
                                    "ultima_actividad": {
                                        "type": "string",
                                        "format": "date-time",
                                        "example": "2025-04-23T14:30:00Z"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Sesión creada exitosamente"
                    },
                    "400": {
                        "description": "Datos inválidos"
                    }
                }
            }
        },
        "/api/sesiones/{id}": {
            "get": {
                "tags": [
                    "Sesiones"
                ],
                "summary": "Obtener una sesión por ID",
                "operationId": "90d2178810d847dbcbe004b9d13f1ede",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la sesión",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Detalles de la sesión"
                    },
                    "404": {
                        "description": "Sesión no encontrada"
                    }
                }
            },
            "put": {
                "tags": [
                    "Sesiones"
                ],
                "summary": "Actualizar una sesión existente",
                "operationId": "259a6fa72dd49782ad18013312f045db",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la sesión",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "usuario_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "agente": {
                                        "type": "string",
                                        "example": "Mozilla/5.0"
                                    },
                                    "ip": {
                                        "type": "string",
                                        "format": "ipv4",
                                        "example": "192.168.1.1"
                                    },
                                    "ultima_actividad": {
                                        "type": "string",
                                        "format": "date-time",
                                        "example": "2025-04-23T14:30:00Z"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Sesión actualizada exitosamente"
                    },
                    "404": {
                        "description": "Sesión no encontrada"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Sesiones"
                ],
                "summary": "Eliminar una sesión",
                "operationId": "a129133fe4358fa4ea8e0fc35d6ea3ab",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la sesión",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Sesión eliminada exitosamente"
                    },
                    "404": {
                        "description": "Sesión no encontrada"
                    }
                }
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    },
...
}